ReadMe
This is a twitter bot that post tweets related to specific users and hashtags. Using Spring and the libraries makes it easy to create a Twitter bot that takes care of the likes of specified hashtags and usernames for you. It manages the Twitter API calls and provides a simple interface. To use the Twitter API the bot needs to authenticate against Twitter